ssh/tailssh: fix incubator tests on macOS arm64
On a system where /bin/true does not exist, the error checked in the
incubator is the wrong error, exec returns different errors depending on
whether it has performed $PATH search.

The code now checks both errors, and also tries /usr/bin/true first, as
that is now and increasingly so, the more common and more portable path.

Updates #16569

Signed-off-by: James Tucker <james@tailscale.com>